+++ custom_model_runner/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 #### [1.4.8] - 2020-12-11
+##### Changes
+- `transform` mode now takes and returns both X and y, although passing y is optional.
 ##### Fixed
 - Force class labels to be strings